# Optimize generate_list_overview()