<!-- What issue does this PR close? -->
Blocked by #6734 -- Rebase this PR after 6734 lands.
Related to #6621 and #6645 which use the same `csv_string()` functions.

<!-- What does this PR achieve? [feature|hotfix|fix|refactor] -->
Optimize `generate_list_overview()` by joining yielded csv strings without building large lists in memory which can make this process unwieldy for users with large reading lists.

### Technical
<!-- What should be noted about the implementation? -->

### Testing
<!-- Steps for reviewer to reproduce/verify what this PR does/fixes. -->
```
Elapsed time (user.get_lists()):         0.016138959 seconds
Elapsed time (generate_list_overview()): 0.002482458 seconds

Elapsed time (user.get_lists()):         0.035167666 seconds
Elapsed time (generate_list_overview()): 0.002955250 seconds

Elapsed time (user.get_lists()):         0.022057500 seconds
Elapsed time (generate_list_overview()): 0.001791541 seconds
```
### Screenshot
<!-- If this PR touches UI, please post evidence (screenshots) of it behaving correctly. -->

### Stakeholders
<!-- @ tag stakeholders of this bug -->


<!-- Attribution Disclaimer: By proposing this pull request, I affirm to have made a best-effort and exercised my discretion to make sure relevant sections of this code which substantially leverage code suggestions, code generation, or code snippets from sources (e.g. Stack Overflow, GitHub) have been annotated with basic attribution so reviewers & contributors may have confidence and access to the correct context to evaluate and use this code. -->